The LSHFunction API

Under construction

This section is currently being developed. If you're interested in helping write this section, feel free to open a pull request; otherwise, please check back later.

API reference

Missing docstring.

Missing docstring for LSHFunction. Check Documenter's build log for details.

LSH.lsh_family — Function

lsh_family(similarity)

Return the default constructor or LSHFunction subtype used to construct a hash function for the similarity function similarity.

The main use of lsh_family is to make it easier to find the documentation for the hash function that's constructed when you call LSHFunction. For instance, if you want to know more about the arguments and keyword parameters that can be given to LSHFunction(inner_prod), you can run

julia> lsh_family(inner_prod)
SignALSH

help?> SignALSH

Examples

julia> lsh_family(cossim)
SimHash

julia> lsh_family(ℓ1)
L1Hash (generic function with 1 method)

See also: LSHFunction