Inner product similarity

SignALSH

LSH.SignALSH — Type

SignALSH(n_hashes::Integer = 1,
         dtype::DataType = Float32,
         maxnorm::Union{Nothing,Real} = nothing,
         m::Integer = 3,
         resize_pow2::Bool = false)

Create a SignALSH hash function for hashing on inner product similarity.

Arguments

n_hashes::Integer (default: 1): the number of hash functions to generate.

Keyword parameters

dtype::DataType (default: Float32): the data type to use in the LSH.SignALSH internals. For performance reasons you should pick dtype to match the type of the data you're hashing.

maxnorm::Union{Nothing,Real} (default: nothing): an upper bound on the $\ell^2$-norm of the data points. Note: this keyword argument must be explicitly specified. If it left unspecified (or set to nothing), SignALSH() will raise an error.

m::Integer (default: 3): parameter m that affects the probability of a hash collision.

resize_pow2::Bool (default: false): affects the way in which the returned LSH.SignALSH resizes to hash inputs of different sizes. If you think you'll be hashing inputs of many different sizes, it's more efficient to set resize_pow2 = true.

Examples

SignALSH is an AsymmetricLSHFunction, and hence hashes must be computed using index_hash and query_hash.

julia> hashfn = SignALSH(12; maxnorm=10);

julia> x = rand(4);

julia> ih = index_hash(hashfn, x); qh = query_hash(hashfn, x);

julia> length(ih) == length(qh) == 12
true

julia> typeof(ih) == typeof(qh) == BitArray{1}
true

You need to explicitly specify the maxnorm keyword parameter when constructing SignALSH, otherwise you will get an error.

julia> hashfn = SignALSH(12)
ERROR: maxnorm must be specified for SignALSH

You'll also get an error if you try to hash a vector that has norm greater than the maxnorm that you specified.

julia> hashfn = SignALSH(; maxnorm=1);

julia> index_hash(hashfn, ones(4))
ERROR: norm 2.0 exceeds maxnorm (1.0)

References

Anshumali Shrivastava and Ping Li. Improved Asymmetric Locality Sensitive Hashing (ALSH) for Maximum Inner Product Search (MIPS). In Proceedings of the Thirty-First Conference on Uncertainty in Artificial Intelligence, UAI'15, page 812–821, Arlington, Virginia, USA, 2015. AUAI Press. 10.5555/3020847.3020931. [https://arxiv.org/abs/1410.5410v2]

See also: inner_prod, ℓ2_norm

MIPSHash

LSH.MIPSHash — Type

MIPSHash(n_hashes::Integer = 1;
         dtype::Datatype = Float32,
         maxnorm::Union{Nothing,Real} = nothing,
         scale::Real = 1,
         m::Integer = 3,
         resize_pow2::Bool = false)

Create a MIPSHash hash function for hashing on inner product similarity.

Arguments

n_hashes::Integer (default: 1): the number of hash functions to generate.

Keyword parameters

dtype::DataType (default: Float32): the data type to use in the LSH.MIPSHash internals. For performance reasons you should pick dtype to match the type of the data you're hashing.

maxnorm::Union{Nothing,Real} (default: nothing): an upper bound on the $\ell^2$-norm of the data points. Note: this keyword argument must be explicitly specified. If it left unspecified (or set to nothing), MIPSHash() will raise an error.

scale::Real (default: 1): parameter that affects the probability of a hash collision. Large values of scale increases hash collision probability (even for inputs with low inner product similarity); small values of scale will decrease hash collision probability.

Examples

MIPSHash is an AsymmetricLSHFunction, and hence hashes must be computed using index_hash and query_hash.

julia> hashfn = MIPSHash(5; maxnorm=10);

julia> x = rand(4);

julia> ih = index_hash(hashfn, x); qh = query_hash(hashfn, x);

julia> length(ih) == length(qh) == 5
true

julia> typeof(ih) == typeof(qh) == Vector{Int32}
true

You need to explicitly specify the maxnorm keyword parameter when constructing MIPSHash, otherwise you will get an error.

julia> hashfn = MIPSHash(5)
ERROR: maxnorm must be specified for MIPSHash

You'll also get an error if you try to hash a vector that has norm greater than the maxnorm that you specified.

julia> hashfn = MIPSHash(; maxnorm=1);

julia> index_hash(hashfn, ones(4))
ERROR: norm 2.0 exceeds maxnorm (1.0)

References

Anshumali Shrivastava and Ping Li. Asymmetric LSH (ALSH) for Sublinear Time Maximum Inner Product Search (MIPS). Proceedings of the 27th International Conference on Neural Information Processing Systems - Volume 2, NIPS'14, page 2321–2329, Cambridge, MA, USA, 2014. MIT Press. 10.5555/2969033.2969086. [https://arxiv.org/abs/1405.5869]

See also: inner_prod, ℓ2_norm

Utilities

LSH.inner_prod — Method

inner_prod(x::AbstractVector, y::AbstractVector)

Computes the $\ell^2$ inner product (dot product)

$\left\langle x, y\right\rangle = \sum_i x_iy_i$

Examples

julia> using LinearAlgebra: dot;

julia> x, y = randn(4), randn(4);

julia> inner_prod(x,y) ≈ dot(x,y)
true