Jaccard similarity

MinHash

LSH.MinHash — Type

MinHash(n_hashes::Integer = 1;
        dtype::DataType = Any,
        symbols::Union{Vector,Set} = Set())

Construct a locality-sensitive hash function for Jaccard similarity.

Arguments

n_hashes::Integer (default: 1): the number of hash functions to generate.

Keyword parameters

dtype::DataType (default: Any): the type of symbols in the sets you're hashing. This is overriden by the data type contained in symbols when symbols is non-empty.

symbols::Union{Vector,Set}: a Vector or Set containing all of the possible elements ("symbols") of the sets that you will be hashing. If left empty, MinHash will instead expand its dictionary when it sees new symbols (at small additional computational expense).

Examples

Construct a hash function to hash sets whose elements are integers between 1 and 50:

julia> hashfn = MinHash(40; symbols = Set(1:50));

julia> n_hashes(hashfn) == 40 && similarity(hashfn) == jaccard
true

julia> hashfn(Set([38, 14, 29, 48, 11]));

julia> hashfn([1, 1, 2, 3, 4]); # You can also hash Vectors

julia> hashfn(Set([100]))
ERROR: Symbol 100 not found

If you aren't sure ahead of time exactly what kinds of elements will be in the sets you're hashing, you can opt not to specify symbols, in which case MinHash will lazily update its hash functions as it encounters new symbols:

julia> hashfn = MinHash();

julia> hashfn(Set([1, 2, 3]));

julia> hashfn(Set(["a", "b", "c"]));

If you don't know what elements you'll encounter, but you know that they'll all be of a specific data type, you can specify the dtype argument for increased efficiency:

julia> hashfn = MinHash(10; dtype = String);

julia> hashfn(Set(["a", "b", "c"]));

References

Broder, A. On the resemblance and containment of documents. Compression and Complexity of Sequences: Proceedings, Positano, Amalfitan Coast, Salerno, Italy, June 11-13, 1997. doi:10.1109/SEQUEN.1997.666900.

See also: jaccard

Utilities

LSH.jaccard — Function

jaccard(A::Set, B::Set) :: Float64

Computes the Jaccard similarity between sets $A$ and $B$, which is defined as

$\text{Jaccard}(A,B) = \frac{\left|A \cap B\right|}{\left|A \cup B\right|}$

Arguments

A::Set, B::Set: the two sets with which to compute Jaccard similarity.

Returns

Float64: the Jaccard similarity between sets A and B, which is between 0 and 1.

Examples

julia> A, B = Set([1, 2, 3]), Set([2, 3, 4]);

julia> jaccard(A,B)
0.5

julia> jaccard(A,B) == length(A ∩ B) / length(A ∪ B)
true

See also: MinHash