$\ell^p$ distance

L1Hash and L2Hash

LSH.L1Hash — Function

L1Hash(
    n_hashes::Integer = 1;
    dtype::DataType = Float32,
    r::Real = 1.0,
    resize_pow2::Bool = false
)

Constructs a locality-sensitive hash for $\ell^1$ distance ($\|x - y\|_1$), defined as

$\|x - y\|_1 = \sum_i |x_i - y_i|$

Arguments

n_hashes::Integer (default: 1): the number of hash functions to generate.

Keyword parameters

dtype::DataType (default: Float32): the data type to use in the L1Hash internals. For performance reasons you should pick dtype to match the type of the data you're hashing.

r::Real (default: 1.0): a positive coefficient whose magnitude influences the collision rate. Larger values of r will increase the collision rate, even for distant points. See references for more information.

resize_pow2::Bool (default: false): affects the way in which the returned L1Hash resizes to hash inputs of different sizes. If you think you'll be hashing inputs of many different sizes, it's more efficient to set resize_pow2 = true.

Examples

Construct an L1Hash with the number of hash functions you want to generate:

julia> hashfn = L1Hash(128);

julia> hashfn.power == 1 &&
       n_hashes(hashfn) == 128 &&
       similarity(hashfn) == ℓ1
true

After creating a hash function, you can compute hashes with hashfn(x):

julia> hashfn = L1Hash(20);

julia> x = rand(4);

julia> hashes = hashfn(x);

References

Datar, Mayur & Indyk, Piotr & Immorlica, Nicole & Mirrokni, Vahab. (2004). Locality-sensitive hashing scheme based on p-stable distributions. Proceedings of the Annual Symposium on Computational Geometry. 10.1145/997817.997857.

See also: ℓ1

LSH.L2Hash — Function

L2Hash(
    n_hashes::Integer = 1;
    dtype::DataType = Float32,
    r::Real = 1.0,
    resize_pow2::Bool = false
)

Constructs a locality-sensitive hash for $\ell^2$ distance ($\|x - y\|_2$), defined as

$\|x - y\|_2 = \left(\sum_i |x_i - y_i|^2\right)^{1/2}$

Arguments

n_hashes::Integer (default: 1): the number of hash functions to generate.

Keyword parameters

dtype::DataType (default: Float32): the data type to use in the L2Hash internals. For performance reasons you should pick dtype to match the type of the data you're hashing.

r::Real (default: 1.0): a positive coefficient whose magnitude influences the collision rate. Larger values of r will increase the collision rate, even for distant points. See references for more information.

resize_pow2::Bool (default: false): affects the way in which the returned L2Hash resizes to hash inputs of different sizes. If you think you'll be hashing inputs of many different sizes, it's more efficient to set resize_pow2 = true.

Examples

Construct an L2Hash with the number of hash functions you want to generate:

julia> hashfn = L2Hash(128);

julia> hashfn.power == 2 &&
       n_hashes(hashfn) == 128 &&
       similarity(hashfn) == ℓ2
true

After creating a hash function, you can compute hashes with hashfn(x):

julia> hashfn = L2Hash(20);

julia> x = rand(4);

julia> hashes = hashfn(x);

References

Datar, Mayur & Indyk, Piotr & Immorlica, Nicole & Mirrokni, Vahab. (2004). Locality-sensitive hashing scheme based on p-stable distributions. Proceedings of the Annual Symposium on Computational Geometry. 10.1145/997817.997857.

See also: ℓ2

Utility functions

LSH.ℓp — Function

ℓp(x::AbstractVector, y::AbstractVector, p::Real=2)
ℓ1(x::AbstractVector, y::AbstractVector)
ℓ2(x::AbstractVector, y::AbstractVector)

Computes the $\ell^p$ distance between a pair of vectors, given by

$\ell^p(x,y) \coloneqq \|x - y\|_p = \left(\sum_i \left|x_i - y_i\right|^p\right)^{1/p}$

ℓ1(x,y) is the same as ℓp(x,y,1), and ℓ2(x,y) is the same as ℓp(x,y,2).

Examples

julia> x = [1, 2, 3];

julia> y = [4, 5, 6];

julia> ℓp(x,y,2) == (abs(1-4)^2 + abs(2-5)^2 + abs(3-6)^2)^(1/2)
true

julia> ℓp(x,y,3) == (abs(1-4)^3 + abs(2-5)^3 + abs(3-6)^3)^(1/3)
true

See also: ℓp_norm, L1Hash, L2Hash

LSH.ℓp_norm — Function

ℓp_norm(x::AbstractVector, p::Real = 2)
ℓ1_norm(x::AbstractVector)
ℓ2_norm(x::AbstractVector)

Compute the $\ell^p$ norm of a point $x$, defined as

$\|x\|_p = \left(\sum_i \left|x_i\right|^p\right)^{1/p}$

Examples

julia> x = randn(4);

julia> ℓp_norm(x, 1) ≈ ℓ1_norm(x) ≈ (map(u -> abs(u)^1, x) |> sum)^(1/1)
true

julia> ℓp_norm(x, 2) ≈ ℓ2_norm(x) ≈ (map(u -> abs(u)^2, x) |> sum)^(1/2)
true

julia> ℓp_norm(x, 3) ≈ (map(u -> abs(u)^3, x) |> sum)^(1/3)
true

See also: ℓp, Lp_norm

LSH.Lp — Method

Lp(x::AbstractVector, y::AbstractVector, p::Real=2)
L1(x::AbstractVector, y::AbstractVector)
L2(x::AbstractVector, y::AbstractVector)

Computes the $ℓ^p$ distance between a pair of vectors $x$ and $y$. Identical to ℓp(x,y,p), ℓ1(x,y), and ℓ2(x,y), respectively.

See also: ℓp

LSH.Lp_norm — Method

Lp_norm(x::AbstractVector, p::Real = 2)
L1_norm(x::AbstractVector)
L2_norm(x::AbstractVector)

Compute the $\ell^p$ norm of a vector $x$. Identical to ℓp_norm(x,p), ℓ1_norm(x), and ℓ2_norm(x), respectively.

See also: ℓp_norm